+++ lib/core/theme/bloc/theme_bloc.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:bloc_concurrency/bloc_concurrency.dart';
-import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:stream_transform/stream_transform.dart';
@@ -18,11 +17,7 @@ EventTransformer<E> throttleDroppable<E>(Duration duration) {
 }
 
 class ThemeBloc extends Bloc<ThemeEvent, ThemeState> {
-  ThemeBloc()
-      : super(ThemeState(
-          theme: FlexThemeData.light(useMaterial3: true),
-          darkTheme: FlexThemeData.dark(useMaterial3: true),
-        )) {
+  ThemeBloc() : super(const ThemeState()) {
     on<ThemeChangeEvent>(
       _themeChangeEvent,
       transformer: throttleDroppable(throttleDuration),
@@ -37,17 +32,21 @@ class ThemeBloc extends Bloc<ThemeEvent, ThemeState> {
       String themeType = prefs.getString('setting_theme_type') ?? 'dark';
       bool useBlackTheme = prefs.getBool('setting_theme_use_black_theme') ?? false;
 
-      // OLED Theme
-      ThemeData? oledThemeData = FlexThemeData.dark(
-        useMaterial3: true,
-        scheme: FlexScheme.deepPurple,
-        darkIsTrueBlack: true,
-      );
+      bool useMaterialYouTheme = prefs.getBool('setting_theme_use_material_you') ?? false;
 
       if (themeType == 'dark') {
-        return emit(state.copyWith(status: ThemeStatus.success, useDarkTheme: true, darkTheme: useBlackTheme ? oledThemeData : ThemeData.dark(useMaterial3: true)));
+        return emit(state.copyWith(
+          status: ThemeStatus.success,
+          useMaterialYouTheme: useMaterialYouTheme,
+          useDarkTheme: true,
+          useBlackTheme: useBlackTheme,
+        ));
       } else {
-        return emit(state.copyWith(status: ThemeStatus.success, useDarkTheme: false));
+        return emit(state.copyWith(
+          status: ThemeStatus.success,
+          useDarkTheme: false,
+          useMaterialYouTheme: useMaterialYouTheme,
+        ));
       }
     } catch (e, s) {
       await Sentry.captureException(e, stackTrace: s);